DateTime + man-page updated

+++ bobcat/datetime/hours.f
@@ -1,4 +1,4 @@
-inline size_t DateTime::hours() const
+inline unsigned DateTime::hours() const
 {
     return d_tm.tm_hour;
 }
