+++ bobcat/datetime/monthdaynr.f
@@ -1,4 +1,4 @@
-inline size_t DateTime::monthDayNr() const
+inline unsigned DateTime::monthDayNr() const
 {
     return d_tm.tm_mday;
 }
