+++ bobcat/datetime/setyear.cc
@@ -1,7 +1,7 @@
 #include "datetime.ih"
 
-void DateTime::setYear(int year)
+void DateTime::setYear(unsigned year)
 {
-                //    s  m  h  md m
-    setFields(TM{ 0, 0, 0, 0, 0, year }, YEAR);
+                //s  m  h  md m
+    setFields(TM{ 0, 0, 0, 0, 0, static_cast<int>(year) }, YEAR);
 }
