+++ bobcat/datetime/weeknr.cc
@@ -1,6 +1,6 @@
 #include "datetime.ih"
 
-size_t DateTime::weekNr() const
+unsigned DateTime::weekNr() const
 {
     unsigned nr = yearDayNr() / 7;
     return nr == 0 ? 52 : nr;
