+++ bobcat/datetime/year.f
@@ -1,4 +1,4 @@
-inline size_t DateTime::year() const
+inline unsigned DateTime::year() const
 {
     return d_tm.tm_year + 1900;
 }
