+++ bobcat/datetime/yeardaynr.f
@@ -1,4 +1,4 @@
-inline size_t DateTime::yearDayNr() const
+inline unsigned DateTime::yearDayNr() const
 {
     return d_tm.tm_yday + 1;
 }
